Fix on WritersCache: delete folder only if it exists

+++ src/core/src/main/java/eu/modernmt/backup/storage/WritersCache.java
@@ -81,7 +81,8 @@ class WritersCache implements Closeable {
             }
         }
 
-        FileUtils.forceDelete(folder);
+        if (folder.exists())
+            FileUtils.forceDelete(folder);
     }
 
     public void flush() throws IOException {
